[feature_extraction] Refactor text/* to text.py
